+++ mycodo/mycodo_flask/method_routes.py
@@ -5,39 +5,33 @@ import datetime
 import time
 import flask_login
 
-from flask import (
-    Blueprint,
-    flash,
-    jsonify,
-    redirect,
-    render_template,
-    request,
-    url_for
-)
+from flask import Blueprint
+from flask import flash
+from flask import jsonify
+from flask import redirect
+from flask import render_template
+from flask import request
+from flask import url_for
+
 from flask_babel import gettext
 from mycodo.mycodo_flask.extensions import db
-# Classes
-from mycodo.databases.models import (
-    DisplayOrder,
-    Method,
-    MethodData,
-    Relay
-)
 
-# Functions
-from mycodo import flaskforms
+from mycodo.databases.models import DisplayOrder
+from mycodo.databases.models import Method
+from mycodo.databases.models import MethodData
+from mycodo.databases.models import Relay
+
+from mycodo.mycodo_flask.forms import forms_method
+
 from mycodo import flaskutils
 from mycodo.mycodo_flask.static_routes import inject_mycodo_version
-from mycodo.utils.system_pi import (
-    csv_to_list_of_int,
-    get_sec,
-    list_to_csv
-)
 
-from mycodo.utils.method import (
-    sine_wave_y_out,
-    bezier_curve_y_out
-)
+from mycodo.utils.system_pi import csv_to_list_of_int
+from mycodo.utils.system_pi import get_sec
+from mycodo.utils.system_pi import list_to_csv
+
+from mycodo.utils.method import sine_wave_y_out
+from mycodo.utils.method import bezier_curve_y_out
 
 logger = logging.getLogger('mycodo.mycodo_flask.methods')
 
@@ -182,7 +176,7 @@ def method_data(method_id):
 @flask_login.login_required
 def method_list():
     """ List all methods on one page with a graph for each """
-    form_create_method = flaskforms.MethodCreate()
+    form_create_method = forms_method.MethodCreate()
 
     method = Method.query.all()
     method_all = MethodData.query.all()
@@ -205,9 +199,9 @@ def method_builder(method_id):
 
     relay = Relay.query.all()
 
-    form_create_method = flaskforms.MethodCreate()
-    form_add_method = flaskforms.MethodAdd()
-    form_mod_method = flaskforms.MethodMod()
+    form_create_method = forms_method.MethodCreate()
+    form_add_method = forms_method.MethodAdd()
+    form_mod_method = forms_method.MethodMod()
 
     # Used in software tests to verify function is executing as admin
     if method_id == '-1':